Se puede eliminar imagen de bd y de storage con if y public_path

+++ app/Http/Controllers/ImagenController.php
@@ -104,9 +104,18 @@ class ImagenController extends Controller
         $imagen->title = $request->title;
 
         if ($request->has('file')) {
-            $destination = $imagen->url;
-            if (File::exists($destination)) {
-                File::delete($destination);
+            $destination = public_path() . $imagen->url;
+            // if (File::exists($destination)) {
+            //     File::delete($destination);
+            // }
+            // if (file_exists(public_path() . $imagen->url)) {
+            // if(File::exists(public_path() . $imagen->url)){
+            //     unlink(public_path() . $imagen->url);
+            // }
+
+            if($imagen->url != '')
+            {
+                unlink(public_path() . '/' . $imagen->url);
             }
 
             $file = $request->file('file');
@@ -133,8 +142,17 @@ class ImagenController extends Controller
 
     public function destroy(Imagen $imagen)
     {
-        $imagen->delete();
-        // eliminar => variable, ok => mensaje
+        // unlink('C:\xampp\htdocs\Mis-Primeros-Auxilitos-Jet\public' . $imagen->url);
+        // if (file_exists(public_path() . $imagen->url)) {
+        //     unlink(public_path() . $imagen->url);
+        //     $imagen->delete();
+        // }
+        if($imagen->url != '')
+        {
+            unlink(public_path() . '/' . $imagen->url);
+            $imagen->delete();
+        }
+        // $imagen->delete();
         return redirect()->route('dashboard.index')->with('eliminar', 'ok');
     }
 }
